+++ sandbox/pages/SandboxBadge.vue
@@ -428,6 +428,7 @@ const handleIconClick = (): void => {
   .horizontal-spacing {
     align-items: flex-end;
     display: flex;
+    flex-wrap: wrap;
     gap: $kui-space-50;
   }
 }