+++ src/components/KCatalog/KCatalog.vue
@@ -54,24 +54,14 @@
     >
       <slot name="error-state">
         <KEmptyState
-          :cta-is-hidden="!errorStateActionMessage || !errorStateActionRoute"
-          :icon="errorStateIcon || ''"
-          :icon-color="errorStateIconColor"
-          :icon-size="errorStateIconSize"
-          is-error
+          :action-button-visible="!!errorStateActionMessage && !!errorStateActionRoute"
+          icon-variant="error"
+          :message="errorStateMessage"
+          :title="errorStateTitle"
         >
-          <template #title>
-            {{ errorStateTitle }}
-          </template>
-
-          <template #message>
-            {{ errorStateMessage }}
-          </template>
-
-          <template #cta>
+          <template #action>
             <KButton
               v-if="errorStateActionMessage"
-              appearance="primary"
               :data-testid="getTestIdString(errorStateActionMessage)"
               :to="errorStateActionRoute ? errorStateActionRoute : undefined"
               @click="$emit('kcatalog-error-cta-clicked')"
@@ -90,20 +80,11 @@
     >
       <slot name="empty-state">
         <KEmptyState
-          :cta-is-hidden="!emptyStateActionMessage || !emptyStateActionRoute"
-          :icon="emptyStateIcon || ''"
-          :icon-color="emptyStateIconColor"
-          :icon-size="emptyStateIconSize"
+          :action-button-visible="!!emptyStateActionMessage && !!emptyStateActionRoute"
+          :message="emptyStateMessage"
+          :title="emptyStateTitle"
         >
-          <template #title>
-            {{ emptyStateTitle }}
-          </template>
-
-          <template #message>
-            {{ emptyStateMessage }}
-          </template>
-
-          <template #cta>
+          <template #action>
             <KButton
               v-if="emptyStateActionMessage"
               :appearance="searchInput ? 'secondary' : 'primary'"
@@ -180,8 +161,8 @@
           :page-sizes="paginationPageSizes"
           :test-mode="!!testMode || undefined"
           :total-count="total"
-          @page-changed="pageChangeHandler"
-          @page-size-changed="pageSizeChangeHandler"
+          @page-change="pageChangeHandler"
+          @page-size-change="pageSizeChangeHandler"
         />
       </div>
     </div>
@@ -199,8 +180,8 @@ import type {
   SwrvStateData,
   CardSize,
   CatalogState,
-  PageChangedData,
-  PageSizeChangedData,
+  PageChangeData,
+  PageSizeChangeData,
 } from '@/types'
 import {
   CardSizeArray,
@@ -542,11 +523,11 @@ const stateData = computed((): SwrvStateData => ({
 }))
 const catalogState = computed((): CatalogState => isCatalogLoading.value ? 'loading' : fetcherError.value ? 'error' : 'success')
 
-const pageChangeHandler = ({ page: newPage }: PageChangedData): void => {
+const pageChangeHandler = ({ page: newPage }: PageChangeData): void => {
   page.value = newPage
 }
 
-const pageSizeChangeHandler = ({ pageSize: newPageSize }: PageSizeChangedData): void => {
+const pageSizeChangeHandler = ({ pageSize: newPageSize }: PageSizeChangeData): void => {
   pageSize.value = newPageSize
   page.value = 1
 }