+++ src/components/KCopy/KCopy.vue
@@ -8,7 +8,7 @@
     </span>
     <div
       class="copy-container"
-      :class="{ 'copy-element': props.truncate || props.badge, 'badge-styles': badge }"
+      :class="{ 'copy-element': truncate || badge, 'badge-styles': badge }"
     >
       <KTooltip
         v-if="format !== 'hidden'"
@@ -18,12 +18,13 @@
         position-fixed
         :text="textTooltipLabel"
       >
-        <span
+        <div
+          ref="copyTextElement"
           class="copy-text"
           :class="{ 'monospace': monospace || !badge }"
         >
           {{ textFormat }}
-        </span>
+        </div>
       </KTooltip>
 
       <KTooltip
@@ -43,6 +44,8 @@
             :size="KUI_ICON_SIZE_30"
             tabindex="0"
             @click.stop="copyIdToClipboard(copyToClipboard)"
+            @keydown.enter="copyIdToClipboard(copyToClipboard)"
+            @keydown.space.prevent="copyIdToClipboard(copyToClipboard)"
           />
         </KClipboardProvider>
       </KTooltip>
@@ -52,12 +55,13 @@
 
 <script setup lang="ts">
 import type { PropType } from 'vue'
-import { computed, ref, watch } from 'vue'
+import { computed, ref, watch, onMounted, onUnmounted } from 'vue'
+import { v4 as uuid4 } from 'uuid'
+import { ResizeObserverHelper } from '@/utilities/resizeObserverHelper'
 import { CopyIcon } from '@kong/icons'
 import KClipboardProvider from '@/components/KClipboardProvider'
 import KTooltip from '@/components/KTooltip/KTooltip.vue'
 import { KUI_ICON_SIZE_30 } from '@kong/design-tokens'
-import { v4 as uuid4 } from 'uuid'
 
 const props = defineProps({
   /**
@@ -130,7 +134,7 @@ const props = defineProps({
    * Number of characters to truncate at
    */
   truncationLimit: {
-    type: Number,
+    type: [Number, String],
     default: 8,
   },
 })
@@ -151,19 +155,17 @@ watch(nonSuccessText, (value: string): void => {
   tooltipText.value = value
 }, { immediate: true })
 
-const truncateLimitText = computed((): string | null => props.truncate ? `${props.text.substring(0, props.truncationLimit) + '...'}` : null)
-
 // Computed for dynamic classes
 const textTooltipClasses = computed((): string => {
   const tooltipWrapperClass = 'copy-tooltip-wrapper' // this selector is referenced in vars.scss - do not update without checking for usage in there first
-  return `${tooltipWrapperClass} ${props.truncate || props.badge ? 'truncate-content' : ''}`
+  return `${tooltipWrapperClass} ${(props.truncate && isTruncated.value) || props.badge ? 'truncate-content' : ''}`
 })
 
 const textFormat = computed(() => {
   if (props.format === 'redacted') {
     return '*****'
   } else if (props.format === 'deleted') {
-    return `*${props.text.substring(0, 5)}`
+    return `*${String(props.text || '').substring(0, 5)}`
   }
 
   // This regex will only remove the quotes if they are the first and last characters of the string (truncateLimitText)
@@ -176,7 +178,7 @@ const textTooltipLabel = computed((): string | undefined => {
   }
 
   // don't show text tooltip if text is redacted or not truncated
-  if (props.format === 'redacted' || !truncateLimitText.value) {
+  if (props.format === 'redacted' || !isTruncated.value) {
     return undefined
   }
 
@@ -209,6 +211,34 @@ const copy = () => {
 defineExpose({
   copy,
 })
+
+const copyTextElement = ref<HTMLDivElement>()
+const resizeObserver = ref<ResizeObserverHelper>()
+const truncateLimitText = computed((): string | null => props.truncate && typeof props.truncationLimit === 'number' ? `${String(props.text || '').substring(0, props.truncationLimit) + '...'}` : null)
+const isTruncated = ref<boolean>(false)
+const setTruncation = (): void => {
+  if (!props.truncate) {
+    return
+  }
+
+  if (props.truncationLimit !== 'auto' && truncateLimitText.value) {
+    isTruncated.value = true
+  } else if (props.truncationLimit === 'auto' && copyTextElement.value) {
+    isTruncated.value = copyTextElement.value.offsetWidth < copyTextElement.value.scrollWidth
+  }
+}
+
+onMounted(() => {
+  resizeObserver.value = ResizeObserverHelper.create(setTruncation)
+
+  resizeObserver.value.observe(copyTextElement.value as HTMLDivElement)
+})
+
+onUnmounted(() => {
+  if (resizeObserver.value) {
+    resizeObserver.value.unobserve(copyTextElement.value as HTMLDivElement)
+  }
+})
 </script>
 
 <style lang="scss" scoped>
@@ -217,6 +247,7 @@ defineExpose({
 .k-copy {
   align-items: center;
   display: flex;
+  max-width: 100%; /** necessary for truncation */
 
   .copy-element {
     align-items: center;
@@ -224,6 +255,10 @@ defineExpose({
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
+
+    .copy-text {
+      @include truncate;
+    }
 
     .truncate-content {
       display: inline-block;
@@ -257,7 +292,7 @@ defineExpose({
     cursor: pointer;
     display: flex;
 
-    .text-icon:not(.k-button .k-copy .text-icon-wrapper .text-icon) {
+    .text-icon:not(.k-button .k-copy .text-icon-wrapper .text-icon):not(.badge-styles .text-icon-wrapper .text-icon) {
       &:hover,
       &:focus {
         // only applies to non-badge as for badge the mixin takes care hover styles
@@ -271,11 +306,6 @@ defineExpose({
     font-size: var(--kui-font-size-20, $kui-font-size-20);
     line-height: var(--kui-line-height-20, $kui-line-height-20);
     margin-right: var(--kui-space-20, $kui-space-20);
-  }
-
-  :deep(.k-popover-content) {
-    font-family: var(--kui-font-family-text, $kui-font-family-text);
-    font-weight: var(--kui-font-weight-regular, $kui-font-weight-regular);
   }
 }
 </style>